+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Deprecations
+
+- [Schema] Column type 'bool' is deprecated — change to 'boolean'
+
 ### Fixed
 
 - Fixed "dependency cycle" warning